renamed src/unlock.ts into src/unlockTypes.ts
