+++ unlock-app/src/__tests__/middlewares/walletMiddleware.test.js
@@ -19,7 +19,7 @@ import { SET_PROVIDER } from '../../actions/provider'
 import { NEW_TRANSACTION } from '../../actions/transaction'
 import { SET_ERROR } from '../../actions/error'
 import { POLLING_INTERVAL } from '../../constants'
-import { TransactionType } from '../../unlock'
+import { TransactionType } from '../../unlockTypes'
 import {
   FATAL_NO_USER_ACCOUNT,
   FATAL_NON_DEPLOYED_CONTRACT,
