+++ unlock-app/src/services/walletService.js
@@ -12,7 +12,7 @@ import {
   FAILED_TO_UPDATE_KEY_PRICE,
   FAILED_TO_WITHDRAW_FROM_LOCK,
 } from '../errors'
-import { TransactionType } from '../unlock'
+import { TransactionType } from '../unlockTypes'
 
 export const keyId = (lock, owner) => [lock, owner].join('-')
